Removed truncating of doubles.
Added function to get x,y coordinates of a specified joint.
Added function to get x,y coordinates of end effector (via previous function).
Added some brief documentation to PaintArm.h functions.
Modified some PaintRobot.cpp functions to make use of above functions.

+++ PaintRobot/Utilities.h
@@ -1,6 +1,6 @@
 struct Point {
-	int x, y;
-	Point(int x, int y) : x(x), y(y) {}
+	double x, y;
+	Point(double x, double y) : x(x), y(y) {}
 };
 typedef struct Point Point;
